leda/mqtt-kuksa-feeder/requirements.txt
paho-mqtt>=2.0.0,<3.0.0
kuksa-client>=0.4.0,<1.0.0
